@@ -5,13 +5,8 @@ version: 1.0.0+1
 publish_to: 'none'
 
 environment:
-  sdk: '>=2.11.0 <3.0.0'
-  flutter: '>=1.25.0'
-
-dependency_overrides:
-  # For DateTime and l10n
-  # TODO: Workaround https://github.com/FirebaseExtended/flutterfire/issues/4342
-  intl: ^0.17.0
+  sdk: '>=2.12.0 <3.0.0'
+  flutter: '>=2.0.0'
 
 dependencies:
   flutter:
@@ -20,44 +15,44 @@ dependencies:
     sdk: flutter
 
   # For Google and Firebase
-  firebase_core: ^0.7.0
-  firebase_auth: ^0.20.0+1
-  google_sign_in: ^4.5.9
-  firebase_crashlytics: ^0.4.0+1
-  firebase_performance: ^0.5.0+1
+  firebase_core: ^1.0.0
+  firebase_auth: ^1.0.0
+  google_sign_in: ^5.0.0
+  firebase_crashlytics: ^1.0.0
+  firebase_performance: ^0.6.0
 
   # For Architecture
-  flutter_hooks: ^0.15.0
-  hooks_riverpod: ^0.12.4
+  flutter_hooks: ^0.16.0
+  hooks_riverpod: ^0.13.0
 
   # For Networking
-  dio: ^3.0.10
-#  dio_http_cache: ^0.2.11
-  dio_firebase_performance: ^0.3.0
+  dio: ^4.0.0-beta4
+  # TODO: dio_firebase_performance does not yet support NNBD.
+  # dio_firebase_performance: ^0.3.0
   # For User-Agent Client Hints
   ua_client_hints: ^1.0.3
 
   # For Model
-  json_serializable: ^3.5.0
-  freezed_annotation: ^0.12.0
+  json_serializable: ^4.0.2
+  freezed_annotation: ^0.14.0
 
   # For DateTime
   intl: ^0.17.0
 
   # For Key-Value local storage
-  shared_preferences: ^0.5.12+4
+  shared_preferences: ^2.0.3
 
   # Convert between Enum and String
-  enum_to_string: ^1.0.14
+  enum_to_string: ^2.0.0-nullsafety.1
 
   # For UIs
-  gap: ^1.2.0
+  gap: ^2.0.0-nullsafety.1
   cupertino_icons: ^1.0.0
-  google_fonts: ^1.1.1
-  flutter_svg: ^0.20.0-nullsafety.3
+  google_fonts: ^2.0.0
+  flutter_svg: ^0.21.0-nullsafety.0
 
   # For Utilities
-  get: ^3.25.4
+  get: ^4.0.0-nullsafety.0
 
 dev_dependencies:
   flutter_test:
@@ -66,14 +61,14 @@ dev_dependencies:
   build_runner: ^1.11.1
 
   # For Model
-  freezed: ^0.12.7
+  freezed: ^0.14.0
 
   # For Analyzer
   effective_dart: ^1.3.0
 
   # For Testing
-  mockito:
-  image_test_utils:
+  mocktail: ^0.0.2-dev.4
+  mocktail_image_network: ^0.0.2-dev.1
 
 flutter_gen:
   integrations:
