@@ -1,10 +1,9 @@
 import 'package:shared_preferences/shared_preferences.dart';
 
 class AppSharedPreferences {
-  SharedPreferences _prefs;
+  SharedPreferences? _prefs;
 
   Future<SharedPreferences> getInstance() async {
-    _prefs ??= await SharedPreferences.getInstance();
-    return _prefs;
+    return _prefs ??= await SharedPreferences.getInstance();
   }
 }
