@@ -1,6 +1,5 @@
 import 'package:dio/adapter.dart';
 import 'package:dio/dio.dart';
-import 'package:dio_firebase_performance/dio_firebase_performance.dart';
 import 'package:flutter/foundation.dart';
 import 'package:ua_client_hints/ua_client_hints.dart';
 
@@ -8,7 +7,7 @@ import '../../constants.dart';
 
 // ignore: prefer_mixin
 class AppDio with DioMixin implements Dio {
-  AppDio._([BaseOptions options]) {
+  AppDio._([BaseOptions? options]) {
     options = BaseOptions(
       baseUrl: Constants.of().endpoint,
       contentType: 'application/json',
@@ -23,7 +22,8 @@ class AppDio with DioMixin implements Dio {
     }));
 
     // Firebase Performance
-    interceptors.add(DioFirebasePerformanceInterceptor());
+    // TODO: dio_firebase_performance does not yet support NNBD.
+    // interceptors.add(DioFirebasePerformanceInterceptor());
 
     if (kDebugMode) {
       // Local Log
