@@ -16,7 +16,7 @@ Future<void> main() async {
   // Crashlytics
   await FirebaseCrashlytics.instance
       .setCrashlyticsCollectionEnabled(kDebugMode);
-  Function originalOnError = FlutterError.onError;
+  Function originalOnError = FlutterError.onError!;
   FlutterError.onError = (errorDetails) async {
     await FirebaseCrashlytics.instance.recordFlutterError(errorDetails);
     originalOnError(errorDetails);
