@@ -8,7 +8,7 @@ void showErrorSnackbar(String title, String message) {
   Get.snackbar(
     title,
     message,
-    backgroundColor: Theme.of(Get.context).errorColor,
+    backgroundColor: Theme.of(Get.context!).errorColor,
     snackPosition: SnackPosition.BOTTOM,
     margin: const EdgeInsets.only(bottom: 8, right: 8, left: 8),
   );
